[Ad Tracker] Fix an incorrect CHECK which caused crashes
This CL removes the CHECK and switches to create
AdAncestorProvenance and NoAdProvenance accordingly.

The theory is that in a non-ad context, the script ID returned from
`IsAdScriptInStackHelper()` could still be null if the ancestry spans
multiple contexts, and an intermediate context is an ad context (leading
to the ID being skipped, e.g. for `bottom_most_async_ad_script_`).

Tests will be added in a follow-up to confirm this theory. The expected
behavior is TBD (i.e., whether we should get the ID even in ad context,
in order to construct the full ancestry chain in such case).

Bug: 421164512
Change-Id: I89284ebab111bfe96c33936a505cf28611c24ae5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6602973
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Yao Xiao <yaoxia@chromium.org>
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1467109}
NOKEYCHECK=True
GitOrigin-RevId: aadc926726500a764b83f25ad2beb15fe3663928