Remove Purpose prefetch header for prefetches and prerender.
Now that prefetches and prerenders are utilizing the Sec-Purpose header for prefetches and prerenders, we will move to remove the legacy Purpose: "prefetch" header that is still currently passed. This will be behind a feature flag/ kill switch to prevent compat issues.

Bug: 420724819
Change-Id: I5d7a8afac76ab08bea760b1af207a0fc612e945a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6594872
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Weizhong Xia <weizhong@google.com>
Commit-Queue: Steven Wei <stevenwei@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1467174}
NOKEYCHECK=True
GitOrigin-RevId: ac284751af237e580fe4369a9e5a26e400122747