Avoid WorkerThread creation if the parent has been terminated
In the previous implementation, we respected `terminate()` for the child
Worker, but we did not see the current parent state.  It brought CHECK
failure when the current worker thread has been terminated while it is
creating the child thread.
To avoid the situation, this CL makes the child thread started only if
the parent thread is live.

Bug: 420993774
Change-Id: I43c8108dd4dff2c76e6a6901f6845dfe2c9420c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6600912
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1467330}
NOKEYCHECK=True
GitOrigin-RevId: c352cb1c99d2f6b9953046407e4bde97f438c151