[Blink] Guard against lack of context in UpdateMemoryUsage()
The context is not necessarily present when this method is called in
DiscardResourceProvider(). This does not come up in tests currently with
`context_` only being dereferenced if the context is WebGL, but it
surfaced in tests in an upcoming change where `context_` will be
dereferenced for Canvas2D as well. It is in any case an actual issue.

Bug: 352263194
Change-Id: I94152c4a4d1a428643b5be6b65bb5fc06a1f2dcf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6596974
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1467341}
NOKEYCHECK=True
GitOrigin-RevId: 49178ec0cc38aa6d8b87f5426f950406917d1430