[Blink] Remove UpdateMemoryUsage() use of ResourceProvider() (part 1)
This CL is the first part of a two-part arc to abstract
HTMLCanvasElement::UpdateMemoryUsage()'s usage of ResourceProvider().
In this part, we abstract its usage of the method to add a "non-GPU
memory buffer" to the memory measurements if the resource provider is
present. To do so, we push that that computation into
ExternallyAllocatedBytesPerPixel().

There is no change in the metrics being computed here.

Bug: 352263194
Change-Id: I1b2260f40b75996e0eee0d66c2882efe8d9b0f43
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6596469
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1467363}
NOKEYCHECK=True
GitOrigin-RevId: 9040a80b31a68b441772567590dae4a324a6cee1