HIC: Add setHitTestRegions scaffolding.
This adds the idl changes as well as plumbs the rects as gfx::RectF
and WeakMember<Element> to HTMLCanvasElement, in preparation for
forwarding hittesting to the canvas elements.

Note that it also does an initial check for IsDrawElementEligible.

We would need to re-verify eligibility when actually doing the
forwarding but I think if the check fails later then we can silently
ignore the particular region.

R=chrishtr@chromium.org

Bug: 421169564
Change-Id: If1f48a41fb4f49a10cfdd23c36fcd7b9c8835753
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6602755
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Vladimir Levin <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1467492}
NOKEYCHECK=True
GitOrigin-RevId: 7616db6075ca98a67eb13b268e23f05ddd8cf342