[LNA] Exempt local network requests from mixed content upgrades
Requests that are known ahead of time to be local network requests
(i.e., private IP literals, .local domains, and fetch() requests with
the targetAddressSpace option set) are already exempted from mixed
content *blocking* (in MixedContentChecker::ShouldBlockFetch()), but the
latter two types are not exempted from mixed content autoupgrading. This
means that they will get upgraded to HTTPS but fail as the local servers
do not support HTTPS. This CL adds a similar exception case to
MixedContentChecker::ShouldAutoupgrade() when the Local Network Access
feature is enabled, a new console message variation to this case, and
unit tests to cover the new behavior.

Bug: 409972641
Change-Id: I425205dd9384e5e0f5498e2a073a6da426c58c74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6606111
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Chris Thompson <cthomp@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1467530}
NOKEYCHECK=True
GitOrigin-RevId: 9a0010ab658e56183e7e64612d5d64dd97fa6732