Metrics for Hidden Input
We added a forward looking feature in IgnoreInputWhileHidden, which
would allow the Renderer to reject input received while hidden. This
broke a11y tools on Windows (crbug.com/407265465).

To understand the scope of this, we are adding a metric to see if other
platforms may have similar complications. This metric is delta from
when the Renderer was hidden until the input arrives.

Bug: 407949600
Change-Id: I4ec4bfc365234ef8da8877b4899fbfb028d93d08
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6603553
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1467560}
NOKEYCHECK=True
GitOrigin-RevId: 2d82043a0101875842a51ad309a783fdc955adde