Translator: Remove CanCreateTranslator
CanCreateTranslator is just TranslationAvailable but without masking.
Using it led to a bug where user activation is not required anytime a
language pair is available despite a site not having created that
language pair before.

Eventually we should also update CanCreateTranslatorResult to
TranslationAvailableResult, but will do this in a follow up later, to
keep this change small and the merge into 138 cleaner.

Bug: 421228289
Change-Id: I6e7318165d496c0fb3aaa1682619148292ca90b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6604677
Reviewed-by: Ayu Ishii <ayui@chromium.org>
Commit-Queue: Nathan Memmott <memmott@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1467582}
NOKEYCHECK=True
GitOrigin-RevId: bcf5e0b87ed5b8e3c2a0d36e9dac6db8d4e580c5