[gap-decorations] Mark GapDecorations flag as experimental
This CL marks the GapDecorations feature flag as experimental, in order
for us to have dev trials on the feature, and we fix test expectation
files accordingly. List of non-obvious fixes:

* computed_style_test.cc: This had to be updated to now take into account that
`ColumnRuleWidth` is now a CSSValueList, so we keep the spirit of the
test, we just make sure to special case the new types.

*flexbox-columns: this test was deleted in this CL because it is testing
that there are no column rules in a flexbox container, which is not true
now with GapDecorations, so the test no longer makes sense.

* column-height-009-ref: as discussed and resolved here
https://github.com/w3c/csswg-drafts/issues/2309 column rules should be
drawn to the height of the container

* column-rule-*-invalid: These are changed because with GapDecorations we do
accept lists of values.

* column-rule-*.html: These were changed to account for the fact that these
rules we no longer treat as CSSKeywordValue and instead are now
CSSStyleValues. With GapDecorations we fail in the set up since now
ruleWidth is a CSSStyleValue rather than a CSSUnitValue, which makes the
test fail before we even get to the asserts. All we can test here now is
whether or not we are a CSSStyleValue. We follow the example of other
properties that are of this type too.

* column-rule invalidation tests rebaseline: the invalidation rect is
bigger now with gap decorations enabled.

* Lastly, we have added interpolation/animation tests as failing to the
test expectations to account for the fact that this does not yet work
with gap decorations. We are aware of this, but we wanted to get the
feature out for dev trials as soon as possible. We have made a tracking
bug for this.

Bug: 357648037, 419066541, 420959039, 421053112

Change-Id: Ie1df2c4987f76008dee872e87db7ca65649b8654
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6514569
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Reviewed-by: Sam Davis Omekara <samomekarajr@microsoft.com>
Commit-Queue: Javier Contreras <javiercon@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1467585}
NOKEYCHECK=True
GitOrigin-RevId: 6520d0ef4b1bcb4977f281e6180131cf9c1f7165