Mark header, footer and page numbering as artifacts during PDF printing
The PDF specification (*) says in section 4.8.2.2.1 that artifacts
that are a side-effect of printing should be marked as such. An
important reason to do that is so they can be filtered out by
accessibility technologies, as they are not relevant to a screen
reader user.

(*) https://opensource.adobe.com/dc-acrobat-sdk-docs/pdfstandards/PDF32000_2008.pdf

Change-Id: I6f1ca832d5ba347919c147218c9edb6908a84d40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6580727
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Chris Harrelson <chrishtr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1467639}
NOKEYCHECK=True
GitOrigin-RevId: 9cf00d9686eefb8f8e93b8d6659499313a2be426