[PEPC] Relax box-shadow restrictions to allow shadows without inset
In order to allow buttons that match material design (which is an often
used tool), we can relax the restriction on box-shadow. However the
inset keyword allows shadows on the inside of the element which is
exactly what we wanted to avoid in the first place. So this CL allows
only box-shadow that don't use the inset keyword.

Fixed: 420864891
Change-Id: I8e17b4590f33d54831fe4168ddf7ebe0ab27e228
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6596955
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Andy Paicu <andypaicu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1468033}
NOKEYCHECK=True
GitOrigin-RevId: 829df277f1daff9b7049bd13c35a3a2310749b7f