Remove incorrect CHECK()
Adding the CHECK() confused counter and counter styles updates. The
LayoutView flag for dirty counters currently survives style and layout
tree updates for ::scroll-markers and page margin boxes.

That is an issue that has been there for a while, which should be
handled somehow, but as a separate task from what introduced this CHECK.

Bug: 421199213, 421837107
Change-Id: I37120c4be2dbd086cd604999001bed9aaa885e6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6610898
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Daniil Sakhapov <sakhapov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1468057}
NOKEYCHECK=True
GitOrigin-RevId: 5a96b580e8d59356614e1bb43b2ac3aa27f32377