[AdTracker] Let IsAdScriptInStackHelper return the frame's creation script, when applicable
When IsAdScriptInStackHelper() returns true, the resulting ad script
ID can be empty if the frame itself is an ad. To improve traceability,
we attempt to return the ad script that created the targeted ad frame.

This change won't address cases where a frame is directly tagged by a
subresource filter without an associated script ancestry. We're
tracking that in bug in (crbug.com/421202278 ).

Bug: 421207631,421202278
Change-Id: Ifd833c183ea706c48094c49fd83361b46b83adb0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6605733
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Yao Xiao <yaoxia@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1468117}
NOKEYCHECK=True
GitOrigin-RevId: 2d8eb71b032afedf876fc237bd6600a626c3252b