Add Blink.HTMLParsing.ResumedByUserTiming UMA
According to the existing metrics, the parser is successfully yielded by
the user timing, but we're not sure how many yielding attempts are
resumed by the user timing, or resumed by the timeout we have set.

This CL adds a boolean histogram to check if the user timing resumed the
parser or not.

Bug: 416543903
Change-Id: I62e199e5caf77ab188c5edc5589b83252f6cadb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6606456
Reviewed-by: Jiacheng Guo <gjc@google.com>
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1468421}
NOKEYCHECK=True
GitOrigin-RevId: ca166ed09cc4f93b89f5b3507dbfce4589652ab8