Bugfix: Avoid calling setRootScriptFilterlistRule() when the rule isn't available
Currently, when the rule isn't available, we call
setRootScriptFilterlistRule() with a null String. However, this still
sets the JavaScript value to an empty string. This CL switches to avoid
calling setRootScriptFilterlistRule() in this scenario, so that the
JavaScript value stays undefined as intended.

Bug: 40194335
Change-Id: I6e3465fb3229dbc63e7549365a2d4912e18b81e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6614278
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Yao Xiao <yaoxia@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1468448}
NOKEYCHECK=True
GitOrigin-RevId: de321537dba4efa40b99ea9a42eb813c7f99a043