Removing kPartitionVisitedLinkDatabase references in code
This flag was used in multi-arm experiments but is currently disabled by
default and will never be launched. It does not control any
functionality that is not currently enabled by
kPartitionVisitedLinkDatabaseWithSelfLinks.

There are small alterations to test suites where this flag was used to
construct various test modes. In history_browsertest and
visitedlink_unittest, I removed the test modes entirely and just enabled
partitioning since there are no longer various levels of partitioning
(just one - with self links).

This is Part 2 of 3 of its removal - we need to remove any references to
the flag before removing the flag itself according to:
https://chromium.googlesource.com/chromium/src/+/HEAD/docs/how_to_add_your_feature_flag.md.

Bug: 421314624
Change-Id: I7b233049982dee591f72a70094a4ff6a0af443f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6604625
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Kyra Seevers <kyraseevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1468510}
NOKEYCHECK=True
GitOrigin-RevId: 3955c7bfd6366b110d27efdc12b0f764ebac41e2