[owners] Remove aaronhk@chromium.org from third_party/blink/renderer/platform/graphics/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/aaronhk@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ic1d8f575848ce036897663ac88427ee5466329a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6576423
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Khushal Sagar <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1468651}
NOKEYCHECK=True
GitOrigin-RevId: 6a8af273447de7cad89ad6d4ea66dd806dc8ec06