[owners] Remove wangxianzhu@chromium.org from third_party/blink/renderer/core/layout/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/wangxianzhu@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I0c247c65f0d79dab9192688041e89e35cebc775e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6618040
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1468798}
NOKEYCHECK=True
GitOrigin-RevId: 129a375ffd11611fbdce460cbdecfeca73cd97d5