[owners] Remove adamk@chromium.org from third_party/blink/web_tests/external/wpt/wasm/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/adamk@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I201f20d71b0a08f19b51a9376f9e58f119da7bb3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6616917
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1468826}
NOKEYCHECK=True
GitOrigin-RevId: e2f1e08c943d45476ca6819c9b1eb089729cec67