[owners] Remove garykac@chromium.org from third_party/blink/web_tests/external/wpt/clipboard-apis/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/garykac@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I1c43cb76c041aac42bc08c741891bf54d600db85
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6618045
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1468848}
NOKEYCHECK=True
GitOrigin-RevId: 560ce72a125dced203eb96caa1f4c2bbc8c5764e