[owners] Remove garykac@chromium.org from third_party/blink/renderer/modules/clipboard/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/garykac@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I6e0742cda9e1bc2763a841891f3f59ccae016468
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6618408
Commit-Queue: Austin Sullivan <asully@chromium.org>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Austin Sullivan <asully@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1468860}
NOKEYCHECK=True
GitOrigin-RevId: c391bcbc8bdbdcb14b4da19a6adc8c63a3f69dd8