[Gap Decorations]: Fix crash when attempting to serialize `initial`
This CL fixes the crash that occurs when rule-properties are set to
`initial`. The crash happened due to the incorrect assumption that with
Gap Decorations, the width property would always be a `CSSValueList`.
However, when the property is parsed using the fast parse path and it's
initial, the parser emits a `CSSInitialValue` (i.e. via the `
ParseCSSWideKeywordValue`).

This change addresses the issue by getting `width_values` as a list,
similar to style and color.

Bug: 421539053
Change-Id: Ia22d66a6239b3cfec3299799b393510dc8463307
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6614377
Commit-Queue: Sam Davis Omekara <samomekarajr@microsoft.com>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1468967}
NOKEYCHECK=True
GitOrigin-RevId: cbeff49a9ec7c3739ce72a3cd95a4dd2d15901c0