Introduce IsSecPurposeForPrefetch
This CL adds a helper method `IsSecPurposeForPrefetch` to check if
`Sec-Purpose` request header set in Chromium is for prefetch. It is
collocated with `Sec-Purpose` header value definitions to help kept
consistent when new header values are added in the future.

Preparation for
https://chromium-review.googlesource.com/c/chromium/src/+/6607183

Bug: 420603369
Change-Id: I413c0615df1be9bb60b33edecc012ddbebffb5dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6616000
Commit-Queue: Hiroshige Hayashizaki <hiroshige@chromium.org>
Reviewed-by: Domenic Denicola <domenic@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469073}
NOKEYCHECK=True
GitOrigin-RevId: c4c68e9232658364ed9e35a3892b5976e1efabcf