[Blink] Make HTMLCanvasElement low-latency checks test Canvas2D directly
The only two context types that support low-latency are WebGL and
Canvas2D, so the checks in HTMLCanvasElement::{Pre, Post}FinalizeFrame()
that are checking "is low-latency && !isWebGL" are effectively checking
"is low-latency && isCanvas2D". This CL moves the checks to the latter
structure to facilitate followup work that will change
GetOrCreateCanvasResourceProvider() calls underneath those checks to
their 2D-specific variant.

Note: One seeming difference here is that we are now explicitly checking
that the context is present. However, for LowLatencyEnabled() to be true
the context must be present anyway [1].

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/html/canvas/html_canvas_element.cc;l=1139-1141?q=HTMLCanvasElement::LowLat&ss=chromium

Bug: 352263194
Change-Id: I1a29d926083c376f2fb753a4d2a0164c7e9f5983
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6616700
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469149}
NOKEYCHECK=True
GitOrigin-RevId: 1291f7e5ac56300d73436acb676c692ddcc1646b