Disallow origin trials for CSS shorthand properties
Execution contexts are not always available for runtime flag checks for
CSS shorthands, which means we cannot always decide if there is an
origin trial enabling a feature or not. We should not allow such
inconsistencies, which may also lead to crashes.

This is an attempt at fixing crash bug 421722988.

Bug: 421722988
Change-Id: Ib905ab57e55e2dba307dcd1e83887e10bc309fca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6619554
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469343}
NOKEYCHECK=True
GitOrigin-RevId: 014ba57cd7be9c42f95f15ccc686c0f008ed161a