Add WaitUntil check for AddRoutes
The addRoutes() method is expected to fail, if it is called outside of
installing. This CL adds WaitUntil() check with the lifetime_resolver.
If the check fails, addRoutes won't be executed. If pass, the install
event will be active until addRoutes complete.

observer is null, can't mock with a unit test, or wpt.

Low-Coverage-Reason: HARD_TO_TEST Unable to test the case when
Bug: 393165254
Change-Id: Ic514e1d3df40965d946fba8cb389f792cf7fe3af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6552597
Commit-Queue: Eain Chen <eainchen@microsoft.com>
Reviewed-by: Liang Zhao <lzhao@microsoft.com>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469370}
NOKEYCHECK=True
GitOrigin-RevId: 68d13a3523498a55cc85ae8e192afad6ebb0c0eb