[vt] Fix raciness with test harness
This test was racy as the teardown conditions for the test were
satisfied before the implications of the exception (skipping the
transition) could be handled. This could leave us with an unhandled
rejection during teardown, or "successfully" completing the test
before it has skipped the transition and had every opportunity to
crash.

Instead, wait for resolution of each of the updateCallbackDone, ready
and finished promises and ensure that all are properly rejected.

Bug: 419851393
Change-Id: Ia66e579201551379da9cd3860d04e687fd347be7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6621104
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469533}
NOKEYCHECK=True
GitOrigin-RevId: 0a6e1f065627b10f01e43290c7a6287f4fc653ea