[soft navs] Make popstate-multiple-backs WPT less racy.
This tweaks the test so that the test awaits the effects of each of the
history.back calls. Also, make the test much clearer in the expected
sequence of the URLs visited, by the soft navigation. It's basically
one.html, two.html, three.html.

This makes it then much easier to adjust the test when submitting
crrev.com/c/6579853: Whereas for now, pushUrl (the expected URL) is
three.html, with crrev.com/c/6579853, it's going to be one.html.

Change-Id: Ibb3187589fe140e52e17789ab6d7b19e38b9ddb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6619730
Commit-Queue: Johannes Henkel <johannes@chromium.org>
Reviewed-by: Michal Mocny <mmocny@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469568}
NOKEYCHECK=True
GitOrigin-RevId: dac5a46ba8778eeadd92518c76ce2c9c0ad0b728