[replaced] Don't attempt to stretch-fit within MinMaxSizesFunc lambda.
This is incorrect, and was a result of us not being about to return
an indefinite size within a MinMaxSizesFunc lambda.
(We can not return kIndefiniteSize as a result of calc-size() related
changes).

Fixes the test which I added testing this behaviour.

Change-Id: I15bbd901a7720c1880249d350f1eb0b29122936e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6618567
Reviewed-by: David Grogan <dgrogan@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469577}
NOKEYCHECK=True
GitOrigin-RevId: c92b480331869b7c1570737a27a0acbb9828edba