Fix resuming the parser while script is running
This is a bug fix for the HTMLParserYieldByUserTiming feature.

Currently, `HTMLDocumentParser` does not expect that the script is not
running when resuming the parser e.g. `ResumeParsingAfterPause()`.

When the feature is enabled, the parser is resumed by either 1)
scheduled timeout or 2) new user timing event via performance.mark().

1) is fine because it's posted as a new task, it doesn't involve script
executions. However, the current implementation of 2) directly call
`NotifyParserResumeByUserTiming()` while running JavaScript, and it hits
DCHECK.

This CL fixes the issue by calling `NotifyParserResumeByUserTiming()` as
a new task in both cases.

Bug: 416543903
Change-Id: I07e28a0a4077581f6128aba03505f1b8efe15230
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6604701
Reviewed-by: Mason Freed <masonf@chromium.org>
Reviewed-by: Minoru Chikamune <chikamune@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469755}
NOKEYCHECK=True
GitOrigin-RevId: 44bfa1b58ab9647de09380a75acf2e72ef510b73