[webview][prefetch] Skip ShouldInterceptRequest & Other Checks for Prefetch Requests
This CL bypasses the following for prefetch requests to remove
latency:

1. The XRW origin trial header check
2. `ShouldInterceptRequest()`

[1] will be deprecated/replaced with a better mechanism in M139,
however the plan is to merge this to M138 to speed up experimentation.

This skipping does not apply to prefetch+prerender requests
which still need to go through all the typical navigation flows
for the time being.

Bug: 420603369, 418159642
Change-Id: I7c71a3a944e7f880c27241ef258406cd484d64d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6607183
Reviewed-by: Peter Beverloo <peter@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Wayne Jackson Jr. <wbjacksonjr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469851}
NOKEYCHECK=True
GitOrigin-RevId: ee18c241d411fd6c4acbc37b059861ab5b209f9c