[SVG] Support width and height as presentation attributes on nested <svg> elements
This CL adds support for the `width` and `height` property
as presentation attributes on the <svg> element.

Firstly in this CL, we enable width and height as presentation
attributes for inner `svg`. Viewport calculation is now based on
the width and height style attributes and not on the width and
height `SVGAnimatedLength` properties.

The above change introduces two cases that needed to be handled
explicitly:

Case 1. Inner `svg` has no presentation attributes defined.
For example, consider:
```html
<svg width="100" height="100">
   <svg>
     <circle cx="50" cy="50" r="40" fill="green" />
   </svg>
 </svg>
```
In the above case, since no attributes is defined on the inner `svg`
parser will never mark the presentation attributes as dirty and
hence they won't be colllected as styles.
However, width and height properties of inner `svg` will still be
computed and hence there is a need to collect these initial
values as styles.

To remedy this, in this case we make use of the fact that styles of
width and height will be `auto` and during their usage in viewport
calculation we will use a length function to resolve them as 100%
instead.

Case 2. When `use` element refer `svg` or `symbol` element,
the styles are actually computed based on the reference element (see
`SVGElement::CustomStyleForLayoutObject`).
When cloned these element will always become an inner `svg` but
they won't follow the usual route of collecting presentation
attributes as styles of the cloned element but instead of the
reference element.
For example, consider:
```html
<svg width="100" height="100">
  <defs>
    <svg id="target">
    <circle cx="50" cy="50" r="40"/>
    </svg>
  </defs>

  <use href="#target" width=50 height=50/>
</svg>
```
According to spec, when width and height presentation attributes
are defined on `use` element, the reference `svg` or `symbol` when
cloned will have those attributes which are defined on `use`.
Since, during styling the presentation attributes which are defined
on the reference element are collected as styles, the
attributes set on cloned element by `use`
(see `void TransferUseWidthAndHeightIfNeeded`) won't be collected
as styles.

To remedy this during presentation attributes calculation of the
reference element we also apply the presentation attributes
available on the cloned element as styles.

This feature is being tracked in:
https://chromestatus.com/feature/5178789386256384

Bug: 40409865
Change-Id: I0d39d7eaaffedf6fc409ed2b3a78fdd4ad4bf5fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6573214
Reviewed-by: Vinay Singh <vinaysingh@microsoft.com>
Commit-Queue: Vinay Singh <vinaysingh@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1469853}
NOKEYCHECK=True
GitOrigin-RevId: f003e3987166d5c5b809c060405998d7e03ea934