ipcz: Synchronize Router::pending_puts_ like other members
Why is |pending_puts_| not protected by |mutex_| like all other members
of Router?

The member moved from Portal into Router when the two classes merged in
crrev.com/1163202. The commit looks almost purely mechanical, so it
likely did not grab any thread safety attention during review. However,
the class Portal at the time already advertised itself as "thread-safe
and manipulated directly by public ipcz API calls" while not
synchronizing access to |pending_puts_|.

Earlier synchronization was _removed_ from Portal in crrev.com/1161995
with this justification:

```
  Remove unnecessary Mutex from Portal. We don't need to synchronize
  on callers' behalf. (-8 bytes)
```

At the time it saved 8 bytes per Portal, but now with Router the lack of
synchronization does not save any memory. Router also claims being
thread safe. So, let's add this synchronization for consistency.

The new locking is short lived, and used only in BeginPut/EndPut, which
I believe is not called at a very high rate. This is good news because
we are unlikely to observe contention in real-world scenarios. But it is
also bad in a sense that this is unlikely to be a reason for flakiness
(see the bug).

A microbenchmark covering this would be good to have. I did not find
one.

Bug: 410594534
Change-Id: I1d91b5bf60561246b15ec57829c091d71470ad3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6619834
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Egor Pasko <pasko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469894}
NOKEYCHECK=True
GitOrigin-RevId: a7a35e37b251765f777ee9391d1096360526ea8a