Improve HTML parser performance in header
This patch replaces a check for the document's body element in
HTMLDocumentParser::ConstructTreeFromToken which was running every time
a token was parsed in the header with a call from the first body element
that gets added to change the parsing mode.

The test case in the bug has tens of megabytes of tags in the head, and
runs significantly faster with this patch. I didn't add it as a test
case because I don't think that having a test file with tens of
megabytes in WPT is worth it.

Hopefully this improves performance of regular pages by some amount as
well.

Fixed: 421318670
Change-Id: Iefa1bfb99a99b5e5eb0e7268eb4a0bf3101baa89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6607631
Reviewed-by: David Baron <dbaron@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1470032}
NOKEYCHECK=True
GitOrigin-RevId: b5e506d2144c45d707d4e87e7f3470bedeeb4b41