Check observer_ before use in RestartForHls
Added a null check for the observer_ WeakPtr before calling
OnHlsManifestDetected() in RestartForHls. Since observer_ is a
WeakPtr, it could be null if the MediaObserver is destroyed.
This prevents a potential crash when HLS media is detected but
the observer is no longer valid.

Change-Id: I42b576336a10aae5bd7a6234a8e047b8986e3d8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6621101
Reviewed-by: Eugene Zemtsov <eugene@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Zeyu Fu <zeyufu@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1470327}
NOKEYCHECK=True
GitOrigin-RevId: 6aaa1886b4ff95b42879664bf6cf5ff71be888a6