[gap-decorations] Fix Flex gaps not painted with only one flex line
This CL fixes a bug where we were not painting gaps when we only had one
flex line.

The way we do this is we strip out the logic for adding the intersection
to the last flex item in each line outside of the logic that handles
each different line (first, middle, last). Essentially, now we check if
we need to add the main gap intersections for the last item in any line
(minus the last line) regardless of which line we are on.

Bug: 357648037
Fixed: 422803236
Change-Id: Id5600cf624f9cad4047e387a0d92040cd1326dab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6625240
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Javier Contreras <javiercon@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1470365}
NOKEYCHECK=True
GitOrigin-RevId: 090ff76f53b38bbb8978f111a654eb7b9cb0ca46