:host selectors incorrectly classified as kNeverMatches
Logical combination pseudos were assumed to never be able to match
against :host in the same compound, which is wrong if the combination
pseudo contains :host. Example: `:is(:host):host` was assumed to never
match because :is() in combination with :host was expected to never
match shadow hosts.

Remove the kNeverMatches optimization for :hosts as it is way too
complex to maintain, adds an extra cost for authors only writing useful
selectors, and most likely has no performance benefit for ill-formed selectors.

Bug: 390453034
Change-Id: I6ac75e025b861d6a4a7b7b9ef459f653ad8c5e20
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6625654
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1470537}
NOKEYCHECK=True
GitOrigin-RevId: d86acf506ba281020538099019a306d47ba27cb2