[soft navs] Improve support for multiple Url changes.
Add support for a Context to have more than a single URL value, by
keeping a reference to the first URL changed, as well as the most recent
URL value observed as attributed to that context.

This helps us carefully do attribution to a specific URL-- for now just
using the first URL consistently.

We also adjust the concept of "most recent context" which will get paint
attribution to now be just the most recent context to update the URL.

Bug: 416705860, 353043684
Change-Id: I8527e32ac1e23a7f0ee292bfa4c77f24fd2a6050
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6579853
Commit-Queue: Michal Mocny <mmocny@chromium.org>
Reviewed-by: Scott Haseley <shaseley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1470592}
NOKEYCHECK=True
GitOrigin-RevId: eb71cbb21b4105873817a3c63b25fc9435f3689b