[gap-decorations] Allow gap decorations to be painted on window resize
This CL fixes a bug where in some cases gap decorations wouldn't be
resized/painted correctly when the window resized.

The issue seemed to be that using the
`ScrollingBackgroundDisplayItemClient` for painting the decorations
would cause this issue, because it seemingly doesn't get marked for
invalidation, which causes the code to use the same drawing recorder
that is used when the window first loads, and doesn't create a new one
when the window is resized since `UseCachedDrawingIfPossible` returns
true for the recorder for that background client. On the flip side,
the DisplayItemClient does get marked for invalidation, so we create
a new recorder for the gap decorations when the window is resized.

Bug: 357648037
Fixed: 422216073
Change-Id: Iabe7ce4fb88e724edd333268ace156299e20c72b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6618134
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Javier Contreras <javiercon@microsoft.com>
Reviewed-by: Sam Davis Omekara <samomekarajr@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1470739}
NOKEYCHECK=True
GitOrigin-RevId: 59e7746d8682528ac4f07dd7b6f3b61712026a0a