Remove VideoTrackRecorder::CodecEnumerator
Most of this code is only used by tests. Other parts are causing
the tests to be flaky since it stores VEA profiles in a static.
AFAICT this is unnecessary caching since these are already cached
in MojoCodecFactory::GetVideoEncodeAcceleratorSupportedProfiles()

The bots don't seem to be hitting this flakiness, but it happens
many times when running these tests locally.

Change-Id: I8aec94cde981d390af130b6b90ce0a3417a9f28d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6626396
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Eugene Zemtsov <eugene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1470741}
NOKEYCHECK=True
GitOrigin-RevId: 5a923b166af7f67caeb4f6cdcf6b0bcee03e42a6