Remove usage of CommitNavigationParams.redirects in the renderer process
The `redirects` field of CommitNavigationParams contains a list of URLs
which are redundant with the `redirect_infos` and `redirect_response`
ones. The list is also off by one, including the start URL and excluding
the URL to commit, while the other two contain all redirects meaning
they do not include the start URL but do include the URL to commit.

The usage of this field on the renderer side does not seem useful and I
think it can be eliminated, so this CL accomplishes this behind a
default enabled flag.

Bug: 422803238
Change-Id: Ia7cd3799cee3d4d46d511f86d89ffd13d57939a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6625816
Reviewed-by: Nate Chapin <japhet@chromium.org>
Commit-Queue: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1470812}
NOKEYCHECK=True
GitOrigin-RevId: 3fc65730592ff510089f6955c09b2881804094d1