Update ServiceWorker classes to run in dedicated worker context
This change updates several Service Worker related classes to support
running in the context of a dedicated worker and not just what they
supported previously of a window. Many of the dependencies on
LocalDOMWindow needed to be updated.

Some code simply CHECKs that the code is running in the context of a
window for example ServiceWorkerContainerHostForClient::Register and
WebServiceWorkerProviderImpl::WebServiceWorkerProviderImpl and can be
modified to also allow the worker context.

Some code uses LocalDOMWindow to get to the ExecutionContext for example
ServiceWorkerContainer is changed to take an ExecutionContext instead of
LocalDOMWindow since that's a base class for both LocalDOMWindow and the
worker contexts. ServiceWorkerContainer also has examples of code that
really should only be for window like the DOMContentLoaded event
handler.

This change is in a series to support running the service worker APIs in
the context of dedicated workers. See https://crbug.com/40364838 for
more details.

Bug: 40364838
Change-Id: Ibc85410f0e2e18356eda38c042bac855d2d29015
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6617354
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Commit-Queue: Dave Risney <david.risney@microsoft.com>
Reviewed-by: Liang Zhao <lzhao@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1470856}
NOKEYCHECK=True
GitOrigin-RevId: 7211caad27a4990ada41d8151e99b5907e9f2da7