Add a new interesttarget test that checks all input modalities
This caught a few bugs and crashes, so this also fixes up some of
the code. In particular, the interest states are more carefully
reset when an element is removed from the DOM.

Note that the long-press portion of the new test crashes due to
an unrelated bug.

Bug: 364669918, 421119389
Change-Id: Iad20f547fe3419f337bb442a7c178e22d7c95023
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6604507
Auto-Submit: Mason Freed <masonf@chromium.org>
Reviewed-by: David Baron <dbaron@chromium.org>
Commit-Queue: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1471052}
NOKEYCHECK=True
GitOrigin-RevId: 9a4edf5c17fb1327a7a55df53157e138ad51e401