Store SVGPropertyBase objects in SVGAnimationElement::values_
Rather than reparsing each pair of values for each interpolation frame,
parse the values once in SVGAnimationElement::UpdateAnimationValues()
and then reuse them until a need to reparse arises (when 'values' or
some state derived from 'values' changes). This also means that
computing the distances for 'paced' animations doesn't need to reparse
values.

This is achieved using a new CalculateValues() that animation elements
implement. This function also gets the task of updating the
end-of-duration value for 'values'-animations - similarly to
CalculateFromAnd{By,To}Values(). This allows eliminating
CalculateToAtEndOfDurationValue() which previously served this purpose
(and did yet another reparse of one of the values).

Rework the 'values'-animation calculations to instead of returning the
string values, return the indices into the 'values' list. Drop the
parsed values cache, replacing it with the pair of indices computed by
the previously mentioned step.

This approach naturally (usually) uses a bit more memory, but has the
benefit of less (re)parsing, and also funnels the parsing of animation
parameters through a single point (UpdateAnimationValues()).

Drop the svg/custom/animation-values-parsing-error.html test, because it
only tests error messages from parsing the basic 'values' attribute and
relies on error reporting to be synchronous. Error reporting for
'values' will now take place as part of animation value validation,
which would make this test a lot more complex.

Bug: 40104599
Change-Id: Ic48333b1590c8d5e672515faa449fbe3ad404cee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6578284
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/main@{#1471229}
NOKEYCHECK=True
GitOrigin-RevId: 41060e0b1158f55fa6e5989324da9f749a03d5f3