[menu] make menulist popover by default
This patch ensures the HTMLMenuListElement always a popover by default
without needing the popover attribute. We accomplish this by ensuring
the popover data is set when the menulist is created.

We also rename HTMLElement::HasPopoverAttribute() to IsPopover().

This patch also adds the commands "toggle-menu", "show-menu" and
"hide-menu" to show/hide the menulist, similarly to the universal
popover commands.

Change-Id: I0f61d8ee67ecd78a6bfd6718333d948e4ceb200f
Bug: 406566432, 422803239
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6612139
Commit-Queue: Di Zhang <dizhangg@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1471311}
NOKEYCHECK=True
GitOrigin-RevId: 660f0cdd7c535cda6c72482e84fa57815aefc87e